refactor abort

+++ controllers/admin-delete.php
@@ -4,8 +4,17 @@
 require("./core/model.php");
 
 $id = $_GET['id'];
+$user_id = 2;
+//!getMovie
+$movie = $fetcher->fetchData("SELECT user_id FROM `tbl_movie` WHERE id=?", [$id], "fetch");
 
+if ($user_id !== $movie['user_id']) {
+  abort(403);
+}
+
+//!DeleteMovie
 $movie = $fetcher->setData('DELETE FROM `tbl_movie` WHERE id=?', [$id]);
 
+//!redirect
 $message = 'successfully Delete movie';
-header('location:' . URL . 'admin?message=' . $message);
+redirect("admin", $message);
